1 - Annual and monthly fee calculation for each record and skip records without average purchase price

+++ src/main/java/com/example/insurance/model/Vehicle.java
@@ -2,72 +2,72 @@ package com.example.insurance.model;
 
 public class Vehicle {
 
-  private String plateNumber;
-  private short firstRegistration;
-  private double purchasePrice;
-  private String producer;
-  private double mileage;
-  private double previousIndemnity;
+    private String plateNumber;
+    private int firstRegistration;
+    private double purchasePrice;
+    private String producer;
+    private double mileage;
+    private double previousIndemnity;
 
-  public Vehicle(String plateNumber,
-                 short firstRegistration,
-                 double purchasePrice,
-                 String producer,
-                 double mileage,
-                 double previousIndemnity) {
-    this.plateNumber = plateNumber;
-    this.firstRegistration = firstRegistration;
-    this.purchasePrice = purchasePrice;
-    this.producer = producer;
-    this.mileage = mileage;
-    this.previousIndemnity = previousIndemnity;
-  }
+    public Vehicle(String plateNumber,
+                   int firstRegistration,
+                   double purchasePrice,
+                   String producer,
+                   double mileage,
+                   double previousIndemnity) {
+        this.plateNumber = plateNumber;
+        this.firstRegistration = firstRegistration;
+        this.purchasePrice = purchasePrice;
+        this.producer = producer;
+        this.mileage = mileage;
+        this.previousIndemnity = previousIndemnity;
+    }
 
-  public String getPlateNumber() {
-    return plateNumber;
-  }
+    public String getPlateNumber() {
+        return plateNumber;
+    }
 
-  public void setPlateNumber(String plateNumber) {
-    this.plateNumber = plateNumber;
-  }
+    public void setPlateNumber(String plateNumber) {
+        this.plateNumber = plateNumber;
+    }
 
-  public short getFirstRegistration() {
-    return firstRegistration;
-  }
+    public int getFirstRegistration() {
+        return firstRegistration;
+    }
 
-  public void setFirstRegistration(short firstRegistration) {
-    this.firstRegistration = firstRegistration;
-  }
+    public void setFirstRegistration(int firstRegistration) {
+        this.firstRegistration = firstRegistration;
+    }
 
-  public double getPurchasePrice() {
-    return purchasePrice;
-  }
+    public double getPurchasePrice() {
+        return purchasePrice;
+    }
 
-  public void setPurchasePrice(double purchasePrice) {
-    this.purchasePrice = purchasePrice;
-  }
+    public void setPurchasePrice(double purchasePrice) {
+        this.purchasePrice = purchasePrice;
+    }
 
-  public String getProducer() {
-    return producer;
-  }
+    public String getProducer() {
+        return producer;
+    }
 
-  public void setProducer(String producer) {
-    this.producer = producer;
-  }
+    public void setProducer(String producer) {
+        this.producer = producer;
+    }
 
-  public double getMileage() {
-    return mileage;
-  }
+    public double getMileage() {
+        return mileage;
+    }
 
-  public void setMileage(double mileage) {
-    this.mileage = mileage;
-  }
+    public void setMileage(double mileage) {
+        this.mileage = mileage;
+    }
 
-  public double getPreviousIndemnity() {
-    return previousIndemnity;
-  }
+    public double getPreviousIndemnity() {
+        return previousIndemnity;
+    }
 
-  public void setPreviousIndemnity(double previousIndemnity) {
-    this.previousIndemnity = previousIndemnity;
-  }
+    public void setPreviousIndemnity(double previousIndemnity) {
+        this.previousIndemnity = previousIndemnity;
+    }
 }
